7.3-stable

+++ usr.bin/mkdep/mkdep.gcc.sh
@@ -1,5 +1,6 @@
 #!/bin/sh -
 #
+#	$OpenBSD: mkdep.gcc.sh,v 1.17 2012/08/30 22:06:43 halex Exp $
 #	$NetBSD: mkdep.gcc.sh,v 1.9 1994/12/23 07:34:59 jtc Exp $
 #
 # Copyright (c) 1991, 1993
@@ -13,11 +14,7 @@
 # 2. Redistributions in binary form must reproduce the above copyright
 #    notice, this list of conditions and the following disclaimer in the
 #    documentation and/or other materials provided with the distribution.
-# 3. All advertising materials mentioning features or use of this software
-#    must display the following acknowledgement:
-#	This product includes software developed by the University of
-#	California, Berkeley and its contributors.
-# 4. Neither the name of the University nor the names of its contributors
+# 3. Neither the name of the University nor the names of its contributors
 #    may be used to endorse or promote products derived from this software
 #    without specific prior written permission.
 #
@@ -36,8 +33,23 @@
 #	@(#)mkdep.gcc.sh	8.1 (Berkeley) 6/6/93
 #
 
-PATH=/bin:/usr/bin:/usr/ucb
-export PATH
+#
+# Scan for a -o option in the arguments and record the filename given.
+# This is needed, since "cc -M -o out" writes to the file "out", not to
+# stdout.
+#
+scanfordasho() {
+	while [ $# != 0 ]
+	do case "$1" in
+		-o)	
+			file="$2"; shift; shift ;;
+		-o*)
+			file="${1#-o}"; shift ;;
+		*)
+			shift ;;
+		esac
+	done
+}
 
 D=.depend			# default dependency file is .depend
 append=0
@@ -66,18 +78,20 @@ while :
 done
 
 if [ $# = 0 ] ; then
-	echo 'usage: mkdep [-p] [-f depend_file] [cc_flags] file ...'
+	echo 'usage: mkdep [-ap] [-f file] [flags] file ...'
 	exit 1
 fi
 
-TMP=/tmp/mkdep$$
+scanfordasho "$@"
 
-trap 'rm -f $TMP ; exit 1' 1 2 3 13 15
+TMP=`mktemp /tmp/mkdep.XXXXXXXXXX` || exit 1
 
-if [ x$pflag = x ]; then
-	gcc -M "$@" | sed -e 's; \./; ;g' > $TMP
+trap 'rm -f $TMP ; trap 2 ; kill -2 $$' 1 2 3 13 15
+
+if [ "x$file" = x ]; then
+	${CC:-cc} -M -w "$@" > $TMP
 else
-	gcc -M "$@" | sed -e 's;\.o :; :;' -e 's; \./; ;g' > $TMP
+	${CC:-cc} -M -w "$@" && cat -- "$file" > $TMP
 fi
 
 if [ $? != 0 ]; then
@@ -86,10 +100,30 @@ if [ $? != 0 ]; then
 	exit 1
 fi
 
+postproc() {
+	in=$1
+	if [ x$pflag = x ]; then
+		sed -e 's; \./; ;g' $in
+	else
+		sed -e 's;\.o[ ]*:; :;' -e 's; \./; ;g' $in
+	fi
+}
+
 if [ $append = 1 ]; then
-	cat $TMP >> $D
-	rm -f $TMP
+	postproc $TMP >> $D
+	if [ $? != 0 ]; then
+		echo 'mkdep: append failed.'
+		rm -f $TMP
+		exit 1
+	fi
 else
-	mv $TMP $D
+	postproc $TMP > $D
+	if [ $? != 0 ]; then
+		echo 'mkdep: rename failed.'
+		rm -f $TMP
+		exit 1
+	fi
 fi
+
+rm -f $TMP
 exit 0
